pyproject.toml
[project]
name = "smart-image-mcp"
version = "0.1.0"
description = "Generic MCP server for image generation supporting AWS Bedrock, OpenAI, and Google Gemini"
readme = "README.md"
requires-python = ">=3.11"
authors = [{name = "Manu Mishra", email = "manumishra@example.com"}]
keywords = ["mcp", "llm", "image-generation", "aws", "openai", "gemini"]
license = { text = "MIT" }
dependencies = [
    "fastmcp>=0.4.1",
    "pillow>=11.1.0",
    "boto3>=1.35.0",
    "openai>=1.0.0",
    "google-genai>=1.7.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
test = ["pytest>=8.0", "pytest-mock>=3.12"]

[project.scripts]
smart-image-mcp = "smart_image_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["smart_image_mcp"]

[tool.pytest.ini_options]
markers = ["integration: marks tests as integration tests"]
